+++ org-code-javabuilder/lib/src/main/java/dev/javabuilder/LocalFileManager.java
@@ -10,7 +10,7 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import org.code.javabuilder.InternalServerError;
-import org.code.javabuilder.Util;
+import org.code.javabuilder.util.FileUtils;
 import org.code.protocol.InternalErrorKey;
 import org.code.protocol.JavabuilderException;
 import org.code.protocol.JavabuilderFileManager;
@@ -62,6 +62,6 @@ public class LocalFileManager implements JavabuilderFileManager {
     // On localhost, we only need to clear the specific temp folder because
     // clearing the entire directory would clear the personal /tmp/ directory
     // in the user's local filesystem.
-    Util.recursivelyClearDirectory(tempFolder.toPath());
+    FileUtils.recursivelyClearDirectory(tempFolder.toPath());
   }
 }
