+++ org-code-javabuilder/lib/src/main/java/org/code/javabuilder/AWSFileManager.java
@@ -14,6 +14,7 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.file.Paths;
 import java.util.Date;
+import org.code.javabuilder.util.FileUtils;
 import org.code.protocol.*;
 
 public class AWSFileManager implements JavabuilderFileManager {
@@ -108,7 +109,7 @@ public class AWSFileManager implements JavabuilderFileManager {
   public void cleanUpTempDirectory(File tempFolder) throws IOException {
     // Delete any leftover contents of the tmp folder from previous lambda invocations
     // We can ignore the tempFolder since we are clearing the entire directory
-    Util.recursivelyClearDirectory(Paths.get(System.getProperty("java.io.tmpdir")));
+    FileUtils.recursivelyClearDirectory(Paths.get(System.getProperty("java.io.tmpdir")));
   }
 
   private String generateKey(String filename) {
