+++ org-code-javabuilder/lib/src/test/java/org/code/javabuilder/FileUtilsTest.java
@@ -2,22 +2,23 @@ package org.code.javabuilder;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import org.code.javabuilder.util.FileUtils;
 import org.junit.jupiter.api.Test;
 
-class FileNameUtilsTest {
+class FileUtilsTest {
 
   @Test
   public void testIsJavaFile() {
     String fileName = "MyClass.java";
-    assertTrue(FileNameUtils.isJavaFile(fileName));
+    assertTrue(FileUtils.isJavaFile(fileName));
 
     fileName = ".java"; // At least one character before extension is required
-    assertFalse(FileNameUtils.isJavaFile(fileName));
+    assertFalse(FileUtils.isJavaFile(fileName));
 
     fileName = "MyClass.txt"; // Java extension required
-    assertFalse(FileNameUtils.isJavaFile(fileName));
+    assertFalse(FileUtils.isJavaFile(fileName));
 
     fileName = "MyClass.Java"; // Java extension must be lowercase
-    assertFalse(FileNameUtils.isJavaFile(fileName));
+    assertFalse(FileUtils.isJavaFile(fileName));
   }
 }
